+++ defender/auto-fund-chainlink-subscription/autotasks/detect-low-balance/index.js
@@ -1,7 +1,7 @@
 import { ethers } from 'ethers';
 import { DefenderRelayProvider } from 'defender-relay-client/lib/ethers';
 import { vrfCoordinatorAddress, subscriptionId, threshold } from '../../subscription-config.dev.yml';
-import ABI from '../../abis/vrfCoordinator.json';
+import ABI from '../../../../abi/vrfCoordinator.json';
 
 //gets called once per block
 export async function handler(event) {
